big fixes and changes to DDS and DMA
Finally fixed low frequencies with DDS, the previous implementation of DDS was incorrect and now the full theoretical frequency range that should be achievable with DDS is possible. This fix makes frequencies such as 0.3 Hz, 0.53 Hz, 1 Hz, 4 Hz, 10 Hz, 100 Hz, and 1 KHz to 500 KHz possible, which is very nice. Unfortunately, correcting DDS made the square wave generation less of a square wave due to capacitor charging/discharging from the low pass filter (needs to be tested more) and also broke how duty cycle works (will need to be looked into more). Using DMA with DDS is also fully fixed, as that was what was done wrong, and works very well.

Square wave generation can possibly be fixed by looking into creating it using sine waves rather than filling the array with HIGH/LOWs based on a given duty cycle.

On the circuit side, I also implemented a voltage follower using an op-amp to separate the function generator and low-pass filter from whatever will be connected to the output of the filter, effectively encapsulating the function generator component of this project entirely from any external circuits that would be powered by the function generator. Previously, without the voltage follower, any loads added to the filter circuit would destroy the signal from the function generator, so the voltage follower was needed to prevent added loads from influencing the characteristics of the filter.